+++ src/main/java/org/tinkoff/notifications/controller/NotificationController.java
@@ -7,6 +7,8 @@ import org.tinkoff.notifications.model.Notification;
 import org.tinkoff.notifications.service.EmployeeService;
 import org.tinkoff.notifications.service.NotificationService;
 
+import static org.tinkoff.notifications.constraint.ApplicationError.NO_NOTIFICATION;
+
 @RestController
 @RequestMapping("/api/notification")
 public class NotificationController {
@@ -29,16 +31,27 @@ public class NotificationController {
     @GetMapping("/get")
     public Notification findById(@RequestParam("notification_id")long notification_id) {
         Notification notification = notificationService.findById(notification_id);
+        if(notification == null) {
+            throw NO_NOTIFICATION.exception(String.format("with id %d", notification_id));
+        }
         return notification;
     }
 
     @PatchMapping("/update")
     public void updateNotification(@RequestBody Notification notification) {
+        Notification notificationCheck = notificationService.findById(notification.getId());
+        if(notificationCheck == null) {
+            throw NO_NOTIFICATION.exception(String.format("with id %d", notification.getId()));
+        }
         notificationService.update(notification);
     }
 
     @DeleteMapping("/delete")
     public void deleteNotification(@RequestBody Notification notification) {
+        Notification notificationCheck = notificationService.findById(notification.getId());
+        if(notificationCheck == null) {
+            throw NO_NOTIFICATION.exception(String.format("with id %d", notification.getId()));
+        }
         notificationService.delete(notification);
     }
 }
