+++ src/main/java/org/tinkoff/notifications/controller/ProjectController.java
@@ -5,6 +5,8 @@ import org.tinkoff.notifications.dto.ProjectDto;
 import org.tinkoff.notifications.model.Project;
 import org.tinkoff.notifications.service.ProjectService;
 
+import static org.tinkoff.notifications.constraint.ApplicationError.NO_PROJECT;
+
 @RestController
 @RequestMapping("/api/project")
 public class ProjectController {
@@ -23,16 +25,27 @@ public class ProjectController {
     @GetMapping("/get")
     public Project findById(@RequestParam("project_id")long project_id) {
         Project project = projectService.findById(project_id);
+        if(project == null) {
+            throw NO_PROJECT.exception(String.format("with id %d", project_id));
+        }
         return project;
     }
 
     @PatchMapping("/update")
     public void updateProject(@RequestBody Project project) {
+        Project projectCheck = projectService.findById(project.getId());
+        if(projectCheck == null) {
+            throw NO_PROJECT.exception(String.format("with id %d", project.getId()));
+        }
         projectService.update(project);
     }
 
     @DeleteMapping("/delete")
     public void deleteProject(@RequestBody Project project) {
+        Project projectCheck = projectService.findById(project.getId());
+        if(projectCheck == null) {
+            throw NO_PROJECT.exception(String.format("with id %d", project.getId()));
+        }
         projectService.delete(project);
     }
 }
